break out summary stats ops from StatsHandler.close
