+++ hive/reporting/handler/stats_handler.py
@@ -12,6 +12,7 @@ from typing import TYPE_CHECKING, List, Dict
 import numpy as np
 
 from hive.reporting.handler.handler import Handler
+from hive.reporting.handler.summary_stats import SummaryStats
 from hive.reporting.report_type import ReportType
 
 if TYPE_CHECKING:
@@ -21,85 +22,21 @@ if TYPE_CHECKING:
 log = logging.getLogger(__name__)
 
 
-@dataclass
-class Stats:
-    state_count: Counter = field(default_factory=lambda: Counter())
-    vkt: Counter = field(default_factory=lambda: Counter())
-
-    requests: int = 0
-    cancelled_requests: int = 0
-
-    mean_final_soc: float = 0
-
-    station_revenue: float = 0
-    fleet_revenue: float = 0
-
-    def compile_stats(self) -> Dict[str, float]:
-        """
-        computes all stats based on values accumulated throughout this run
-        :return: a dictionary with stat values by key
-        """
-        requests_served_percent = 1 - (self.cancelled_requests / self.requests) if self.requests > 0 else 0
-        total_state_count = sum(self.state_count.values())
-        total_vkt = sum(self.vkt.values())
-        vehicle_state_output = {}
-        vehicle_states_observed = set(self.state_count.keys()).union(self.vkt.keys())
-        for v in vehicle_states_observed:
-            observed_pct = self.state_count.get(v) / total_state_count if self.state_count.get(v) else 0
-            vkt = self.vkt.get(v, 0)
-            data = {
-                "observed_percent": observed_pct,
-                "vkt": vkt
-            }
-            vehicle_state_output.update({v: data})
-
-        output = {
-            "mean_final_soc": self.mean_final_soc,
-            "requests_served_percent": requests_served_percent,
-            "vehicle_state": vehicle_state_output,
-            "total_vkt": total_vkt,
-            "station_revenue_dollars": self.station_revenue,
-            "fleet_revenue_dollars": self.fleet_revenue
-        }
-
-        return output
-
-    def log(self):
-        log.info(f"{self.mean_final_soc * 100:.2f} % \t Mean Final SOC".expandtabs(15))
-        requests_served_percent = (1 - (self.cancelled_requests / self.requests)) * 100 if self.requests > 0 else 0
-        log.info(
-            f"{requests_served_percent:.2f} % \t Requests Served"
-                .expandtabs(15)
-        )
-
-        total_state_count = sum(self.state_count.values())
-        for s, v in self.state_count.items():
-            log.info(f"{round(v / total_state_count * 100, 2)} % \t Time in State {s}".expandtabs(15))
-
-        total_vkt = sum(self.vkt.values())
-        log.info(f"{total_vkt:.2f} km \t Total Kilometers Traveled".expandtabs(15))
-        for s, v in self.vkt.items():
-            log.info(f"{v:.2f} km \t Kilometers Traveled in State {s}".expandtabs(15))
-
-        log.info(f"$ {self.station_revenue:.2f} \t Station Revenue".expandtabs(15))
-        log.info(f"$ {self.fleet_revenue:.2f} \t Fleet Revenue".expandtabs(15))
-
-
 class StatsHandler(Handler):
     """
     The StatsHandler compiles various simulation statistics and stores them.
     """
 
     def __init__(self):
-        self.stats = Stats()
+        self.stats = SummaryStats()
 
-    def get_stats(self) -> Dict:
+    def get_stats(self, rp: RunnerPayload) -> Dict:
         """
         special output specifically for the StatsHandler which produces the
         summary file output
         :return: the compiled stats for this simulation run
         """
-        return self.stats.compile_stats()
+        return self.stats.compile_stats(rp)
 
     def handle(self, reports: List[Report], runner_payload: RunnerPayload):
         """
@@ -143,27 +80,8 @@ class StatsHandler(Handler):
 
         :return:
         """
-        sim_state = runner_payload.s
-        env = runner_payload.e
-
-        self.stats.mean_final_soc = np.mean([
-            env.mechatronics.get(v.mechatronics_id).fuel_source_soc(v) for v in sim_state.vehicles.values()
-        ])
-
-        self.stats.station_revenue = reduce(
-            lambda income, station: income + station.balance,
-            sim_state.stations.values(),
-            0.0
-        )
-
-        self.stats.fleet_revenue = reduce(
-            lambda income, vehicle: income + vehicle.balance,
-            sim_state.vehicles.values(),
-            0.0
-        )
-
+        output = self.stats.compile_stats(runner_payload)
         self.stats.log()
-        output = self.stats.compile_stats()
         output_path = Path(runner_payload.e.config.scenario_output_directory).joinpath("summary_stats.json")
         with output_path.open(mode="w") as f:
             json.dump(output, f, indent=4)
